first part of phase 7
